C201032__session-2/fastapi-news/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy
databases
pydantic
requests-html
lxml_html_clean
python-dotenv
mysql-connector-python
